Remove activity class
