+++ app/src/main/AndroidManifest.xml
@@ -18,9 +18,6 @@
             android:name=".ui.filter.FilterCardsActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.collectionexport.ExportCollectionActivity"
-            android:exported="false" />
-        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Bootcamp.NoActionBar">
