+++ app/src/main/java/com/github/sdp/tarjetakuna/MainActivity.kt
@@ -41,7 +41,8 @@ class MainActivity : AppCompatActivity() {
                 R.id.nav_home,
                 R.id.nav_browser,
                 R.id.nav_scanner,
-                R.id.nav_webapi),
+                R.id.nav_webapi,
+                R.id.nav_exportcollection),
             drawerLayout)
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
