+++ app/src/main/java/com/github/sdp/tarjetakuna/ui/authentication/SignOutActivity.kt
@@ -9,7 +9,6 @@ import androidx.appcompat.app.AppCompatActivity
 import com.github.sdp.tarjetakuna.MainActivity
 import com.github.sdp.tarjetakuna.R
 import com.github.sdp.tarjetakuna.database.UserCardsActivity
-import com.github.sdp.tarjetakuna.ui.collectionexport.ExportCollectionActivity
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 
@@ -42,12 +41,6 @@ class SignOutActivity: AppCompatActivity() {
         val configureCollectionButton  = findViewById<Button>(R.id.manage_collection_button)
         configureCollectionButton.setOnClickListener {
             val intent = Intent(this, UserCardsActivity::class.java)
-            startActivity(intent)
-        }
-
-        val exportCollectionButton  = findViewById<Button>(R.id.export_collection_button)
-        exportCollectionButton.setOnClickListener {
-            val intent = Intent(this, ExportCollectionActivity::class.java)
             startActivity(intent)
         }
     }
