--- app/src/main/java/com/github/sdp/tarjetakuna/ui/collectionexport/ExportCollectionActivity.kt
@@ -1,113 +1,0 @@
-package com.github.sdp.tarjetakuna.ui.collectionexport
-
-import android.content.Intent
-import android.net.Uri
-import android.os.Bundle
-import android.os.Environment
-import android.view.View
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.FileProvider
-import com.github.sdp.tarjetakuna.R
-import com.github.sdp.tarjetakuna.model.*
-import com.github.sdp.tarjetakuna.ui.authentication.SignOutActivity
-import org.apache.poi.hssf.usermodel.HSSFWorkbook
-import org.apache.poi.ss.usermodel.Row
-import org.apache.poi.ss.usermodel.Sheet
-import org.apache.poi.ss.usermodel.Workbook
-import org.apache.poi.ss.util.CellUtil
-import java.io.File
-import java.io.FileOutputStream
-import java.io.IOException
-
-/**
- * This activity is responsible to create an excel file containing the user's collection, permitting the user to export it
- */
-class ExportCollectionActivity : AppCompatActivity() {
-
-    private val fileProviderAuthority = "com.github.sdp.tarjetakuna.fileprovider"
-
-    //This is a test collection, it will be replaced later by the user's collection
-    private val testCollection = listOf(
-        MagicCard("MagicCard", "A beautiful card", MagicLayout.Normal, 7, "{5}{W}{W}", MagicSet("MT15", "Magic 2015"), 56, "https://img.scryfall.com/cards/large/front/1/2/12345678.jpg?1562567890"),
-        MagicCard("BestMagicCard", "An even more beautiful card", MagicLayout.Normal, 7, "{7}{W}{W}", MagicSet("MT15", "Magic 2015"), 56, "https://img.scryfall.com/cards/large/front/1/2/12345678.jpg?1562567890"),
-    )
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_export_collection)
-
-        //create a new excel file
-        val fileName = "MyMagicCollection.xls"
-        val filePath =
-            getExternalFilesDir(Environment.DIRECTORY_DOCUMENTS)?.absolutePath + "/" + fileName
-        val excelFile = File(filePath)
-
-        //write the collection to the excel file
-        writeDataToExcel(excelFile, testCollection)
-
-        //open the excel file
-        openExcelFile(excelFile)
-
-        val button = findViewById<View>(R.id.button_back_to_sign_out_page)
-        button.setOnClickListener {
-            val intent = Intent(this, SignOutActivity::class.java)
-            startActivity(intent)
-        }
-
-    }
-
-    /**
-     * Write the collection to the excel file
-     */
-    private fun writeDataToExcel(excelFile: File, collection: List<MagicCard>) {
-        try {
-            val workbook: Workbook = HSSFWorkbook()
-            val sheet: Sheet = workbook.createSheet("NewSheet")
-
-            // Create header row
-            val headerRow: Row = sheet.createRow(0)
-            CellUtil.createCell(headerRow, 0, "Name")
-            CellUtil.createCell(headerRow, 1, "Description")
-            CellUtil.createCell(headerRow, 2, "Mana Cost")
-            CellUtil.createCell(headerRow, 3, "Set")
-            CellUtil.createCell(headerRow, 4, "Image URL")
-
-            // Create data rows
-            for (i in collection.indices) {
-                val dataRow: Row = sheet.createRow(i + 1)
-                CellUtil.createCell(dataRow, 0, collection[i].name)
-                CellUtil.createCell(dataRow, 1, collection[i].text)
-                CellUtil.createCell(dataRow, 2, collection[i].manaCost)
-                CellUtil.createCell(dataRow, 3, collection[i].set.name)
-                CellUtil.createCell(dataRow, 4, collection[i].imageUrl)
-            }
-
-            // Save workbook to file
-            val fileOutputStream = FileOutputStream(excelFile)
-            workbook.write(fileOutputStream)
-            fileOutputStream.close()
-        } catch (e: IOException) {
-            Toast.makeText(this, "Failed to write data to Excel file", Toast.LENGTH_SHORT).show()
-            e.printStackTrace()
-        }
-    }
-
-    /**
-     * Open the view to share the excel file
-     */
-    private fun openExcelFile(excelFile: File) {
-        val uri: Uri = FileProvider.getUriForFile(this, fileProviderAuthority, excelFile)
-
-        val intent = Intent(Intent.ACTION_SEND)
-        intent.type = "application/excel"
-        intent.putExtra(Intent.EXTRA_STREAM, uri)
-        intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-
-        try {
-            startActivity(intent)
-        } catch (e: Exception) {
-            Toast.makeText(this, "No app found to share Excel file", Toast.LENGTH_SHORT).show()
-        }
-    }
-}
